requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
numpy==1.26.2
Werkzeug==3.0.1
gunicorn==21.2.0